+++ omq-libzmq/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Enable `zstd+tcp://` transport through the C API.
+
 ## [0.5.9] - 2026-07-31
 
 ### Changed